+++ test/fixtures/comma-dangle.js
@@ -1,0 +1,19 @@
+/* eslint-disable no-unused-vars, comma-spacing,
+   node/no-unsupported-features/es-syntax */
+
+const fun = () => {
+  /* */
+}
+let arr, obj
+
+arr = [1, 2]
+arr = [1, 2]
+
+obj = { a: 1, b: 2 }
+obj = {
+  a: 1,
+  b: 2,
+}
+
+fun(1, 2)
+fun(1, 2)