+++ test/fixtures/react1.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const ReactTest = () => (
+  <div className="none" data-url="#">
+    {'foo'}
+  </div>
+)